requirements.txt
nextcord
aiohttp
python-dotenv
cloudscraper
flask
Pillow
scipy
lxml
python-dateutil